backend/requirements.txt
django-cors-headers
markdown
django-filter
Django==4.1.2
channels==4.0.0
channels-redis==4.0.0
celery==5.2.7
single-beat
django-celery-results
django-celery-beat
djangorestframework==3.14.0
djangorestframework-dataclasses
drf-spectacular[sidecar]
django_nextjs
ipython
django-jazzmin
psycopg2-binary
uvicorn[standard]
whitenoise
jinja2
psycopg2-binary
django-model-utils
requests
django-prometheus
bs4
ddrr
django-revproxy
openai